Added new DAO libs

+++ src/main/java/by/epam/labproject/createmypc/service/CrudService.java
@@ -11,7 +11,7 @@ public interface CrudService <T> {
 
     void update(T domain) throws ServiceException;
 
-    void delete(Integer id) throws ServiceException;
+    void delete(Long id) throws ServiceException;
 
     Iterable<T> findAll();
 }
